feat: `stats` add a `fast-precision` feature
the current way we calculate precision is using ryu to convert a float to a string, and then looking at the fractional part and taking its length.

Though ryu is faster than stdlib, its still allocating.

The fast-precision feature turns on an alternative way to estimate precision using bit manipulation.

It's non-allocating and faster.

However, we get different results from the two different methods, so we leave it to the user.